echo middleware: support being called with raw context
The echo middleware handles interacting with iam-runtime using an echo context.
Results are then checked and a proper echo error is returned.

However when the user has a standard context.Context instead of an echo.Context,
they can no longer take advantage of the echo middleware's ability to return
proper echo errors.

One scenario where this may come up is when adding tracing spans to your methods.
The first thing you do is create a new span which returns a standard context.Context.
In order to take advantage of the echo middleware while keeping the trace context,
the user would have to replace the echo.Context's request with one with the updated context.

This adds unnecessary work for the user as the echo middleware simply unwraps the
request context before processing it anyways.
So a simple solution which simplifies usability is to simply expose a function which
can act on a context.Context while still handling echo errors properly.

Signed-off-by: Mike Mason <mimason@equinix.com>